Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/evaluation-error-4-expected.txt

CONSOLE MESSAGE: line 2: [object Object]

FAIL Test that exceptions during evaluation lead to error events on window, and that exceptions are remembered.
 assert_array_equals: lengths differ, expected 6 got 2
